Refactor encryption functionality into new class Separate util functions into a new file
