+++ src/mesh/blockchain/BlockchainHandler.cpp
@@ -1,5 +1,6 @@
 #if !MESHTASTIC_EXCLUDE_WEBSERVER
 #include "BlockchainHandler.h"
+#include "BlockchainUtils.h"
 #include "FSCommon.h"
 #include "HTTPClient.h"
 #include "WiFi.h"
@@ -17,75 +18,11 @@
 #include <memory>
 #include <sstream>
 
-// Redefine strptime in your source to avoid IRAM issue
-char *strptime(const char *str, const char *format, struct tm *tm)
-{
-    if (sscanf(str, format, &tm->tm_year, &tm->tm_mon, &tm->tm_mday, &tm->tm_hour, &tm->tm_min, &tm->tm_sec) == 6) {
-        tm->tm_year -= 1900; // Adjust year to be relative to 1900
-        tm->tm_mon -= 1;     // Adjust month to be 0-based
-        return (char *)(str + strlen(str));
-    }
-    return NULL;
-}
-
-String getCurrentTimestamp()
-{
-    // Get current time
-    std::time_t now = std::time(nullptr);
-    std::tm *now_tm = std::gmtime(&now);
-
-    // Use stringstream to format the time
-    std::ostringstream oss;
-    oss << std::put_time(now_tm, "%Y-%m-%d %H:%M:%S");
-    oss << " UTC";
-
-    return String(oss.str().c_str());
-}
-
-void logLongString(const String &str, size_t chunkSize = 50)
-{
-    size_t len = str.length();
-    if (len <= chunkSize) {
-        LOG_INFO("%s\n", str.c_str());
-    } else {
-        size_t i = 0;
-        while (i < len) {
-            size_t end = std::min(i + chunkSize, len);
-            LOG_INFO("%s\n", str.substring(i, end).c_str());
-            i = end;
-        }
-    }
-}
-
-// Function to convert enum to string
-std::string blockchainStatusToString(BlockchainStatus status)
-{
-    switch (status) {
-    case BlockchainStatus::OK:
-        return "OK";
-    case BlockchainStatus::NO_WIFI:
-        return "NO_WIFI";
-    case BlockchainStatus::HTTP_ERROR:
-        return "HTTP_ERROR";
-    case BlockchainStatus::EMPTY_RESPONSE:
-        return "EMPTY_RESPONSE";
-    case BlockchainStatus::PARSING_ERROR:
-        return "PARSING_ERROR";
-    case BlockchainStatus::NODE_NOT_FOUND:
-        return "NODE_NOT_FOUND";
-    case BlockchainStatus::READY:
-        return "READY";
-    case BlockchainStatus::NOT_DUE:
-        return "NOT_DUE";
-    default:
-        return "UNKNOWN_STATUS";
-    }
-}
-
 BlockchainHandler::BlockchainHandler(const std::string public_key, const std::string private_key)
     : public_key_(public_key), private_key_(private_key)
 {
     kda_server_ = "http://kda.crankk.org/chainweb/0.0/mainnet01/chain/19/pact/api/v1/";
+    encryptionHandler_ = std::unique_ptr<EncryptionHandler>(new EncryptionHandler());
 }
 
 bool BlockchainHandler::isWalletConfigValid()
@@ -95,7 +32,6 @@ bool BlockchainHandler::isWalletConfigValid()
 
 int32_t BlockchainHandler::performNodeSync(HttpAPI *webAPI)
 {
-
     LOG_INFO("\nWallet public key: %s\n", public_key_.data());
     LOG_INFO("\nWallet private key: %s\n", private_key_.data());
 
@@ -113,7 +49,7 @@ int32_t BlockchainHandler::performNodeSync(HttpAPI *webAPI)
     if (status == BlockchainStatus::READY) { // node exists, due for sending
         uint32_t packetId = generatePacketId();
         String secret_hex = String(packetId, HEX);
-        String secret = encrypt(director_pubkeyd_, secret_hex.c_str());
+        String secret = encryptPayload(secret_hex.c_str());
         status = executeBlockchainCommand("send", "(free.mesh03.update-sent \"" + secret + "\")");
         if (status == BlockchainStatus::OK) {
             // Only send the radio beacon if the update-sent command is successful
@@ -133,71 +69,6 @@ int32_t BlockchainHandler::performNodeSync(HttpAPI *webAPI)
     auto newHeap = memGet.getFreeHeap();
     LOG_INFO("Free heap: %d\n", newHeap);
     return 300000; // Every 5 minutes. That should be enough for previous txn to be complete
-}
-
-struct HashVector {
-    const char *name;
-    const char *data;
-    uint8_t hash[HASH_SIZE];
-};
-
-uint8_t *BlockchainHandler::Binhash(BLAKE2b *hash, const struct HashVector *test)
-{
-    size_t size = strlen(test->data);
-    static uint8_t value[HASH_SIZE];
-
-    hash->reset(32);
-    hash->update(test->data, size);
-    hash->finalize(value, sizeof(value));
-
-    return value;
-}
-
-String BlockchainHandler::KDAhash(BLAKE2b *hash, const struct HashVector *test)
-{
-    size_t size = strlen(test->data);
-    uint8_t value[HASH_SIZE];
-
-    hash->reset(32);
-    hash->update(test->data, size);
-    hash->finalize(value, sizeof(value));
-
-    auto inputLength = sizeof(value);
-    char output[base64::encodeLength(inputLength)];
-    base64::encode(value, inputLength, output);
-    String hashString = String(output);
-    hashString.replace("+", "-");
-    hashString.replace("/", "_");
-    hashString.replace("=", "");
-    return hashString;
-}
-
-void BlockchainHandler::HexToBytes(const std::string &hex, char *out)
-{
-    for (unsigned int i = 0; i < hex.length(); i += 2) {
-        std::string byteString = hex.substr(i, 2);
-        char byte = (char)strtol(byteString.c_str(), NULL, 16);
-        out[i / 2] = byte;
-    }
-}
-
-String BlockchainHandler::generateSignature(const uint8_t *hashBin)
-{
-    char publicKey[32];
-    HexToBytes(public_key_, publicKey);
-
-    char privateKey[32];
-    HexToBytes(private_key_, privateKey);
-
-    uint8_t signature[64];
-    Ed25519::sign(signature, (uint8_t *)privateKey, (uint8_t *)publicKey, hashBin, HASH_SIZE);
-
-    // Convert bytes to hex string
-    String signHex = "";
-    for (uint8_t i = 0; i < sizeof(signature); i++) {
-        signHex += String(signature[i] < 16 ? "0" : "") + String(signature[i], HEX);
-    }
-    return signHex;
 }
 
 JSONObject BlockchainHandler::createCommandObject(const String &command)
@@ -239,15 +110,15 @@ JSONObject BlockchainHandler::preparePostObject(const JSONObject &cmdObject, con
     postObject["cmd"] = new JSONValue(cmdString.c_str());
     HashVector vector{"Test1", cmdString.c_str()};
 
-    uint8_t *hashBin = Binhash(&blake2b_, &vector);
-    String hash = KDAhash(&blake2b_, &vector);
+    uint8_t *hashBin = encryptionHandler_->Binhash(&vector);
+    String hash = encryptionHandler_->KDAhash(&vector);
     // LOG_INFO("\n%s\n", hash.c_str());
 
     // for (uint8_t i = 0; i < HASH_SIZE; i++) {
     //     LOG_INFO("%d,%d\n", i, hashBin[i]);
     // }
 
-    String signHex = generateSignature(hashBin);
+    String signHex = encryptionHandler_->generateSignature(public_key_, private_key_, hashBin);
     postObject["hash"] = new JSONValue(hash.c_str());
     JSONArray sigs;
     JSONObject sigObject{{"sig", new JSONValue(signHex.c_str())}};
@@ -339,211 +210,36 @@ BlockchainStatus BlockchainHandler::executeBlockchainCommand(String commandType,
     }
 }
 
-// Function to add PKCS7 padding
-void add_pkcs7_padding(std::vector<unsigned char>& data, size_t block_size) {
-    size_t padding_size = block_size - (data.size() % block_size);
-    data.insert(data.end(), padding_size, static_cast<unsigned char>(padding_size));
-}
-
-void EvpKDF(const unsigned char *password, size_t password_len,
-            const unsigned char *salt, size_t salt_len,
-            unsigned char *pOutKey, size_t key_size,
-            unsigned char *pOutIV, size_t iv_size,
-            mbedtls_md_type_t md_type, int iterations) {
-    mbedtls_md_context_t md_ctx;
-    mbedtls_md_init(&md_ctx);
-
-    if (mbedtls_md_setup(&md_ctx, mbedtls_md_info_from_type(md_type), 1)!= 0) {
-        fprintf(stderr, "Failed to setup md context\n");
-        return;
-    }
-
-    size_t block_size = mbedtls_md_get_size(mbedtls_md_info_from_type(md_type));
-    std::vector<unsigned char> block(block_size);
-    std::vector<unsigned char> derivedKey;
-
-    size_t total_size = key_size + iv_size;
-
-    while (derivedKey.size() < total_size) {
-        mbedtls_md_starts(&md_ctx);
-        if (!derivedKey.empty()) {
-            mbedtls_md_update(&md_ctx, block.data(), block.size());
-        }
-        mbedtls_md_update(&md_ctx, password, password_len);
-        mbedtls_md_update(&md_ctx, salt, salt_len); // Incorporate salt
-        mbedtls_md_finish(&md_ctx, block.data()); // Finalize the block
-
-        for (int i = 1; i < iterations; i++) {
-            mbedtls_md_starts(&md_ctx);
-            mbedtls_md_update(&md_ctx, block.data(), block_size);
-            mbedtls_md_finish(&md_ctx, block.data()); // Correctly finalize the block once per iteration
-        }
-
-        derivedKey.insert(derivedKey.end(), block.begin(), block.end());
-    }
-
-    mbedtls_md_free(&md_ctx);
-
-    // Ensure the derivedKey is long enough
-    if (derivedKey.size() > total_size) {
-        derivedKey.resize(total_size);
-    }
-
-    memcpy(pOutKey, derivedKey.data(), key_size);
-    memcpy(pOutIV, derivedKey.data() + key_size, iv_size);
-}
-
-// Function to log key and IV in hexadecimal format
-void logEncryptionInfo(const unsigned char* key, size_t keySize, const unsigned char* iv, size_t ivSize, const unsigned char* salt, size_t saltSize, const std::vector<unsigned char>& encryptedData) {
-    std::string keyHex;
-    for (size_t i = 0; i < keySize; ++i) {
-        char hex[3]; // Temporary buffer for each byte
-        sprintf(hex, "%02x", key[i]);
-        keyHex += hex; // Append the hex string to the result
-    }
-    LOG_INFO("Derived KEY: %s\n", keyHex.c_str());
-
-    std::string ivHex;
-    for (size_t i = 0; i < ivSize; ++i) {
-        char hex[3]; // Temporary buffer for each byte
-        sprintf(hex, "%02x", iv[i]);
-        ivHex += hex; // Append the hex string to the result
-    }
-    LOG_INFO("Derived IV: %s\n", ivHex.c_str());
-
-    // Convert and log the salt in hexadecimal format
-    std::string saltHex;
-    for (size_t i = 0; i < saltSize; ++i) {
-        char hex[3]; // Temporary buffer for each byte
-        sprintf(hex, "%02x", salt[i]);
-        saltHex += hex; // Append the hex string to the result
-    }
-    LOG_INFO("Salt: %s\n", saltHex.c_str());
-
-    std::string encryptedDataHex;
-    for (unsigned char byte : encryptedData) {
-        char hex[3]; // Temporary buffer for each byte
-        sprintf(hex, "%02x", byte);
-        encryptedDataHex += hex; // Append the hex string to the result
-    }
-    LOG_INFO("Encrypted Data: %s\n", encryptedDataHex.c_str());
-}
-
-String BlockchainHandler::encrypt(const std::string &base64PublicKey, const std::string &payload) {
-    // Initialize mbedTLS structures
-    mbedtls_aes_context aes;
-    mbedtls_pk_context pk;
-    mbedtls_entropy_context entropy;
-    mbedtls_ctr_drbg_context ctr_drbg;
-    mbedtls_rsa_context *rsa;
-
-    // Initialize and seed the random number generator
-    mbedtls_ctr_drbg_init(&ctr_drbg);
-    mbedtls_entropy_init(&entropy);
-    mbedtls_aes_init(&aes);
-    mbedtls_pk_init(&pk);
-
-    // Seed the random number generator
-    if (mbedtls_ctr_drbg_seed(&ctr_drbg, mbedtls_entropy_func, &entropy, nullptr, 0) != 0) {
-        LOG_ERROR("Failed to initialize RNG\n");
+String BlockchainHandler::encryptPayload(const std::string &payload) {
+    if (!encryptionHandler_) {
+        LOG_ERROR("Encryption handler is not initialized. Encryption failed.\n");
         return "";
     }
+    return encryptionHandler_->encrypt(director_pubkeyd_, payload);
+}
 
-    // Decode Base64 public key to binary
-    size_t decodedKeyLength = base64::decodeLength(base64PublicKey.c_str());
-    std::vector<uint8_t> binaryKey(decodedKeyLength);
-    base64::decode(base64PublicKey.c_str(), binaryKey.data());
-    std::string decodedKey(binaryKey.begin(), binaryKey.end());
-
-    // Convert std::string to const unsigned char*
-    const unsigned char* publicKey = reinterpret_cast<const unsigned char*>(decodedKey.c_str());
-    size_t publicKeyLen = decodedKey.size() + 1;
-    if (mbedtls_pk_parse_public_key(&pk, publicKey, publicKeyLen) != 0) {
-        LOG_ERROR("Failed to parse public key\n");
-        return "";
+// Function to convert enum to string
+std::string BlockchainHandler::blockchainStatusToString(BlockchainStatus status)
+{
+    switch (status) {
+    case BlockchainStatus::OK:
+        return "OK";
+    case BlockchainStatus::NO_WIFI:
+        return "NO_WIFI";
+    case BlockchainStatus::HTTP_ERROR:
+        return "HTTP_ERROR";
+    case BlockchainStatus::EMPTY_RESPONSE:
+        return "EMPTY_RESPONSE";
+    case BlockchainStatus::PARSING_ERROR:
+        return "PARSING_ERROR";
+    case BlockchainStatus::NODE_NOT_FOUND:
+        return "NODE_NOT_FOUND";
+    case BlockchainStatus::READY:
+        return "READY";
+    case BlockchainStatus::NOT_DUE:
+        return "NOT_DUE";
+    default:
+        return "UNKNOWN_STATUS";
     }
-
-    rsa = mbedtls_pk_rsa(pk);
-    mbedtls_rsa_set_padding(rsa, MBEDTLS_RSA_PKCS_V21, MBEDTLS_MD_SHA256);
-
-    // Generate a random 16-byte symmetric key
-    unsigned char aesKey[16];
-    mbedtls_ctr_drbg_random(&ctr_drbg, aesKey, sizeof(aesKey));
-
-    // Convert symmetric key to hex-encoded string
-    char symKeyHex[33];
-    for (int i = 0; i < 16; i++) {
-        sprintf(&symKeyHex[i * 2], "%02x", aesKey[i]);
-    }
-    symKeyHex[32] = '\0';
-
-    // Encrypt the symmetric key using RSA-OAEP with SHA-256
-    unsigned char buffer[512];
-    size_t olen;
-    if (mbedtls_rsa_rsaes_oaep_encrypt(rsa, mbedtls_ctr_drbg_random, &ctr_drbg, MBEDTLS_RSA_PUBLIC, nullptr, 0, sizeof(symKeyHex), reinterpret_cast<const unsigned char*>(symKeyHex), buffer) != 0) {
-        LOG_ERROR("RSA encryption failed\n");
-        return "";
-    }
-    olen = mbedtls_rsa_get_len(rsa);
-
-    // Generate a random 8-byte salt
-    unsigned char salt[8];
-    mbedtls_ctr_drbg_random(&ctr_drbg, salt, sizeof(salt));
-    unsigned char derivedKey[32], derivedIV[16];
-    std::string symKeyString = std::string(symKeyHex);
-    LOG_INFO("KEY: %s\n", symKeyString.c_str());
-
-    // IMPORTANT: SymKeyHex must be 33 bytes (counting the null terminator), otherwise key derivation will differ from CryptoJS
-    EvpKDF(reinterpret_cast<const unsigned char*>(symKeyHex), 33, salt, 8, derivedKey, 32, derivedIV, 16, MBEDTLS_MD_MD5, 1);
-
-    // Set key length to 256 bits
-    mbedtls_aes_setkey_enc(&aes, derivedKey, 256);  // Using 256-bit encryption key
-
-    // Add PKCS7 padding to the payload
-    std::vector<unsigned char> paddedPayload(payload.begin(), payload.end());
-    add_pkcs7_padding(paddedPayload, 16);
-
-    // Encrypt the payload using AES-CBC
-    std::vector<unsigned char> encryptedData(paddedPayload.size());
-    unsigned char derivedIVCopy[16];
-    memcpy(derivedIVCopy, derivedIV, sizeof(derivedIVCopy));
-    mbedtls_aes_crypt_cbc(&aes, MBEDTLS_AES_ENCRYPT, paddedPayload.size(), derivedIVCopy, paddedPayload.data(), encryptedData.data());
-
-    logEncryptionInfo(derivedKey, 32, derivedIV, 16, salt, 8, encryptedData);
-
-    // Combine prefix, salt and encryptedData
-    std::vector<unsigned char> combinedData;
-    const char* prefix = "Salted__";
-    combinedData.insert(combinedData.end(), prefix, prefix + 8);
-    combinedData.insert(combinedData.end(), salt, salt + sizeof(salt));
-    combinedData.insert(combinedData.end(), encryptedData.begin(), encryptedData.end());
-
-    // Base64 encode the encrypted AES key
-    size_t encryptedKeyBase64Len = base64::encodeLength(olen);
-    std::vector<char> encryptedKeyBase64(encryptedKeyBase64Len);
-    base64::encode(buffer, olen, encryptedKeyBase64.data());
-
-    // Base64 encode the combined data
-    size_t combinedDataBase64Len = base64::encodeLength(combinedData.size());
-    std::vector<char> combinedDataBase64(combinedDataBase64Len);
-    base64::encode(combinedData.data(), combinedData.size(), combinedDataBase64.data());
-
-    std::string encryptedCombinedDataStr(combinedDataBase64.begin(), combinedDataBase64.end());
-    std::string encryptedKeyStr(encryptedKeyBase64.begin(), encryptedKeyBase64.end());
-
-    // Concatenate the encoded strings with a delimiter
-    String result = String(encryptedCombinedDataStr.c_str()) + ";;;;;" + String(encryptedKeyStr.c_str());
-    LOG_INFO("+++++++++ Encrypted combined Data Base64: %s\n", encryptedCombinedDataStr.c_str());
-    LOG_INFO("+++++++++ Encrypted SymKey Base64: %s\n", encryptedKeyStr.c_str());
-    LOG_INFO("+++++++++ RESULT OF ENCRYPT!!!: %s\n", result.c_str());
-    logLongString(result);
-
-    // Cleanup
-    mbedtls_aes_free(&aes);
-    mbedtls_pk_free(&pk);
-    mbedtls_ctr_drbg_free(&ctr_drbg);
-    mbedtls_entropy_free(&entropy);
-
-    return result;
 }
 #endif
